gemma/gemma-sglang/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pillow>=10.0.0
aiohttp>=3.9.0
numpy>=1.24.0
requests>=2.31.0
python-multipart>=0.0.6